backend/requirements.txt
# ggNet Backend Requirements
# Python 3.11+

# FastAPI and web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0         # SQLite async driver
alembic==1.13.1           # Database migrations

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
psutil==5.9.8             # System metrics
jinja2==3.1.3             # Template rendering for configs
python-dotenv==1.0.0      # Environment variables
aiofiles==23.2.1          # Async file operations

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0             # Async HTTP client for testing

